+++ lib/qr_code_dart_scan.dart
@@ -1,12 +1,12 @@
 library qr_code_dart_scan;
 
 export 'package:camera/camera.dart';
-export 'package:qr_code_dart_decoder/qr_code_dart_decoder.dart' show CroppingStrategy;
+export 'package:qr_code_dart_decoder/qr_code_dart_decoder.dart'
+    show BarcodeFormat, CroppingStrategy, ScanResult;
 
 export 'src/decoder/qr_code_dart_scan_decoder.dart';
-export 'src/models/barcode_format.dart';
-export 'src/models/scan_result.dart';
 export 'src/qr_code_dart_scan_controller.dart';
 export 'src/qr_code_dart_scan_view.dart';
 export 'src/util/image_decode_orientation.dart';
 export 'src/util/qr_code_dart_scan_resolution_preset.dart';
+export 'src/util/scan_point_extensions.dart';